% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mirror_desc}
\alias{mirror_desc}
\title{Mirror Descent (Exponentiated Gradient Descent) for Synthetic Control}
\usage{
mirror_desc(X0_scaled, X1_scaled, V, niter = 10000L, rel_tol = 1e-08)
}
\arguments{
\item{X0_scaled}{The scaled matrix of controls}

\item{X1_scaled}{The scaled matrix (column vector) for the treated values}

\item{V}{The diagonal matrix for the weights for each variable}

\item{niter}{The number of interations for the algorithm. Defauls to
10,000}

\item{rel_tol}{The stopping tolerance. Defaults to 1e-8}
}
\description{
Mirror Descent (Exponentiated Gradient Descent) for Synthetic Control.
}
\details{
This algorithm is from Aethy et al. (2017; Matrix Completion). This
function will minimize with respect to w

(X0 * w - X1)'(X0 * w - X1)

The matrix X where X0 is the controls and X1 is the treated vector.
Before passing X1 and X0 to this function, X needs to be scaled
to have zero rowmeans, variance 1, and normalized by the largest value.
}

