% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MirrorDescSynth_main.R
\docType{package}
\name{MirrorDescSynth-package}
\alias{MirrorDescSynth}
\alias{MirrorDescSynth-package}
\title{Mirror Descent for Synthetic Control}
\description{
MirrorDescSynth implements the mirror descent (Exponentiated
Gradient Descent) for the synthetic control optimization
}
\details{
The only function you're likely to need from \pkg{roxygen2} is
\code{\link{roxygenize}}. Otherwise refer to the vignettes to see
how to format the documentation.
}

